[Back] refacto api calls into utils/apiCalls
