+++ Back/test/loggerTest.js
@@ -1,11 +1,7 @@
 const fileTools = require('./utils/fileTools');
+const calls = require('./utils/apiCalls');
 var assert = require('assert');
-var request = require('supertest');
-var User = require('../models/userModel');
-var app = require('../app');
-var agent = request.agent(app);
 
-var response;
 const dummyName = "Dummy Foo";
 const dummyEmail = "Dummy.Foo@asylum.io";
 const dummyPassword = "MyP4ZZVV0RDEZ";
@@ -13,68 +9,12 @@ const dummy = {name: dummyName, password: dummyPassword, email: dummyEmail};
 const dummyWithToken = {token: process.env.TOKEN_TEST};
 const dummyWithoutPassword = {name: dummyName, email: dummyEmail};
 
-/// Test Helpers
-const getPoints = () => {
-    return new Promise((resolve, error) => {
-        agent.get('/points/')
-        .end((err, res) => {
-            if(err) error(err);
-            else resolve();
-        });
-    });
-}
-
-/// Test Helpers
-const regiterDummy = (data) => {
-    return new Promise((resolve, error) => {
-        agent.post('/users/register')
-        .send(data)
-        .end((err, res) => {
-            if(err) error(err);
-            else resolve();
-        });
-    });
-}
-
-const loginDummy = (data) => {
-    return new Promise((resolve, error) => {
-        agent.post('/users/login')
-        .send(data)
-        .end((err, res) => {
-            response = res;
-            if(err) error(err);
-            else resolve();
-        });
-    });
-}
-
-const loginDummyWithToken = (token) => {
-    return new Promise((resolve, error) => {
-        agent.post('/users/login')
-        .set('Authorization', token)
-        .end((err, res) => {
-            if(err) { done(err);}
-            else {
-                response = res;
-                resolve();
-            }
-        });
-    });
-}
-
-const deleteDummy = (data) =>
-    new Promise((resolve, error) => {
-        User.deleteOne({name: data.name, email: data.email}, (err) => {
-            if(err) error(err);
-            else resolve();
-        });
-    });
-
-/// TESTS
 describe('Logger tests', () => {
     describe('Log #GET request without body', () => {
         before((done) => {
-            getPoints().then(done).catch(done);
+            calls.getPoints()
+            .then(_ => done()) // Ignore response parameter 
+            .catch(done); // Will call done with error 
         });
 
         it('Logger expects to write into logs files', (done) => {
@@ -121,13 +61,9 @@ describe('Logger tests', () => {
     describe('Log #POST request with not well formed body', () => {
         describe('Log #POST register request without password', () => {
             before((done) => {
-                regiterDummy(dummyWithoutPassword)
-                .then(res => {
-                    done();
-                })
-                .catch(err => {
-                    done();
-                })
+                calls.regiterDummy(dummyWithoutPassword)
+                .then(_ => done()) // Ignore response parameter
+                .catch(_ => done()); // Avoid error handling and tests fails
             });
 
             it('Logger expects to write last register error', (done) => {
@@ -143,11 +79,9 @@ describe('Logger tests', () => {
                     .catch(done);
             });
 
-            after(
-                (done) => {
-                    deleteDummy(dummy).then(done).catch(done);
-                }
-            );
+            after((done) => {
+                calls.deleteDummy(dummy).then(done).catch(done);
+            });
         });
     });
 
@@ -155,11 +89,12 @@ describe('Logger tests', () => {
         describe('Log #POST login request with password', () => {
             before(
                 (done) => {
-                    regiterDummy(dummy)
+                    calls.regiterDummy(dummy)
                     .then(
-                        _ => {
-                            loginDummy(dummy).then(done).catch(done);
-                        }
+                        _ => 
+                            calls.loginDummy(dummy)
+                            .then(_ => done())
+                            .catch(done)
                     )
                     .catch(done)
                 }
@@ -177,21 +112,20 @@ describe('Logger tests', () => {
                     .catch(done);
             });
 
-            after(
-                (done) => {
-                    deleteDummy(dummy).then(done).catch(done);
-                }
-            );
+            after((done) => {
+                calls.deleteDummy(dummy).then(done).catch(done);
+            });
         });
 
         describe('Log #POST login request with token', () => {
             before(
                 (done) => {
-                    regiterDummy(dummy)
+                    calls.regiterDummy(dummy)
                     .then(
-                        _ => {
-                            loginDummyWithToken(dummyWithToken).then(done).catch(done);
-                        }
+                        _ => 
+                            calls.loginDummyWithToken(dummyWithToken)
+                            .then(_ => done())
+                            .catch(done)
                     )
                     .catch(done)
                 }
@@ -208,11 +142,9 @@ describe('Logger tests', () => {
                     .catch(done);
             });
 
-            after(
-                (done) => {
-                    deleteDummy(dummy).then(done).catch(done);
-                }
-            );
+            after((done) => {
+                calls.deleteDummy(dummy).then(done).catch(done);
+            });
         });
     });
